Portfolio_with_auth_removed/requirements.txt
asgiref==3.2.10
boto==2.49.0
boto3==1.14.47
botocore==1.17.47
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.8
django-ses==1.0.3
django-storages==1.9.1
djangorestframework==3.11.1
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
jmespath==0.10.0
lxml==4.6.2
numpy==1.19.0
oauth2client==4.1.3
pandas==1.0.5
praw==7.1.0
prawcore==1.4.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
twilio==6.45.3
update-checker==0.17
urllib3==1.25.9
websocket-client==0.57.0
whitenoise==5.2.0
